requirements.txt
appnope==0.1.0
astroid==2.1.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==2.1.5
django-extensions==2.1.4
djangorestframework==3.9.0
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.2.2
pytz==2018.9
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.11.0
